document AlphaVantageResponse, AlphaVantageWebClientService, GlobalQuote and StockController
